Add support for working a Signed OCI (SOCI) image
A signed OCI (SOCI) is a metadata OCI Image which includes:

- an install manifest which specifies what container images to install,
  some container config, and where the container is pulled from.
- a sigature of the install manifest by a private key
- the public key of the signing key

With these three elements, a SOCI aware system can verify the containers
it will install are from a trusted source (assuming you trust the
signer) and will install and run a certain way without being tampered.

Implement a SOCIRef api to inspect and get the compoents.

Verify and display information about the SOCI
  $ ocidist soci inspect ocidist://host:port/path/to/install:version

Extract the SOCI artifacts
  $ ocidist soci get ocidist://host:port/path/to/install:version
  {
      "install": <install manifest>,
      "pubkeycrt": <cert here>,
      "signature": {
          "encoding": "base64",
          "data": <b64(signature)>,
      },
  }

Signed-off-by: Ryan Harper <rharper@woxford.com>